requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.1
django-filter==21.1
Pillow==9.0.0
sqlparse==0.4.2
